pyproject.toml
[tool.poetry]
name = "taskpy"
version = "0.1.0"
description = ""
authors = ["Vicente Fernandes <reis.vicente27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.6"
djangorestframework = "^3.14.0"
markdown = "^3.4.4"
django-filter = "^23.2"
pep8 = "^1.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
